refactor(cms): move `topicsOrderBySpeechCount` typedefs and resolver to extend-graphql-schemas/topics-order-by-speech-count.ts
