+++ packages/cms/extend-graphql-schemas/index.ts
@@ -3,44 +3,18 @@ import type { GraphQLSchema } from 'graphql'
 import { mergeSchemas } from '@graphql-tools/schema'
 // custom sql
 import {
-  getTopicsSql,
-  getTop5LegislatorSql,
-} from '../custom-sql/topicsOrderBySpeechCount'
-import {
   recentSpeechTopicStatsTypeDefs,
   recentSpeechTopicStatsResolvers,
 } from './recent-speech-topic-stats'
+import {
+  topicsOrderBySpeechCountTypeDefs,
+  topicsOrderBySpeechCountResolver,
+} from './topics-order-by-speech-count'
 import { getLegislatorsSql } from '../custom-sql/topNTopicsOfLegislators'
 // lodash
 import { groupBy } from 'lodash'
 const _ = {
   groupBy,
-}
-
-type TopicWithSpeechCount = {
-  id: number
-  title: string
-  slug: string
-  speechCount: bigint
-  legislatorCount: bigint
-}
-type LegislatorForTopic = {
-  id: number
-  slug: string
-  name: string
-  count?: number
-  party?: number
-  imageLink?: string
-  image?: {
-    imageFile: {
-      url: string
-    }
-  }
-}
-type LegislatorForTopicRaw = LegislatorForTopic & {
-  topicId: number
-  imageId?: number
-  imageExtension?: string
 }
 
 type TopicForLegislator = {
@@ -58,30 +32,6 @@ const extendGraphqlSchema = (baseSchema: GraphQLSchema) => {
     schemas: [baseSchema],
     typeDefs: [
       `
-      """ custom type for topicsOrderBySpeechCount """
-      type ImageFileWithUrlOnly {
-        url: String!
-      }
-      type CustomImage {
-        imageFile: ImageFileWithUrlOnly!
-      }
-      type LegislatorForTopic {
-        id: Int!
-        name: String!
-        party: Int
-        count: Int
-        slug: String!
-        imageLink: String
-        image: CustomImage
-      }
-      type TopicWithSpeechCountAndLegislators {
-        title: String!
-        slug: String!
-        speechCount: Int!
-        legislatorCount: Int!
-        legislators: [LegislatorForTopic]
-      }
-      
       """ custom type for topNTopicsOfLegislators """
       type TopicWithSpeechCount {
         title: String!
@@ -95,78 +45,16 @@ const extendGraphqlSchema = (baseSchema: GraphQLSchema) => {
       }
 
       type Query {
-        """ Get topics order by speech count desc """
-        topicsOrderBySpeechCount(meetingId: Int!, sessionIds: [Int] = [], partyIds: [Int] = [], take: Int = 10, skip: Int = 0): [TopicWithSpeechCountAndLegislators]
         """ Get top N topic of a legislator order by speech cound desc """
         topNTopicsOfLegislators(legislatorIds: [Int!]!, meetingId: Int!, sessionIds: [Int] = [], take: Int = 10): [LegislatorWIthTopTopics]
       }
     `,
       recentSpeechTopicStatsTypeDefs,
+      topicsOrderBySpeechCountTypeDefs,
     ],
     resolvers: [
       {
         Query: {
-          topicsOrderBySpeechCount: async (
-            _root,
-            { meetingId, sessionIds, partyIds, take, skip },
-            context: Context
-          ) => {
-            const topics = await context.prisma.$queryRaw<
-              TopicWithSpeechCount[]
-            >(getTopicsSql({ meetingId, sessionIds, partyIds, take, skip }))
-            if (!topics || topics.length === 0) {
-              return []
-            }
-            const topicIds = topics.map(({ id }) => id)
-            const top5legislators = await context.prisma.$queryRaw<
-              LegislatorForTopicRaw[]
-            >(
-              getTop5LegislatorSql({
-                meetingId,
-                sessionIds,
-                partyIds,
-                topicIds,
-              })
-            )
-            type GroupedLegislators = {
-              [topicId: number]: Array<LegislatorForTopic>
-            }
-            const top5legislatorsGrouped = top5legislators.reduce(
-              (
-                acc: GroupedLegislators,
-                { count, topicId, imageId, imageExtension, ...rest }
-              ) => {
-                if (!acc[topicId]) {
-                  acc[topicId] = []
-                }
-                let image
-                if (imageId && imageExtension) {
-                  image = {
-                    imageFile: {
-                      url: `/images/${imageId}.${imageExtension}`, // todo: use config in keystone.config
-                    },
-                  }
-                }
-                const res = {
-                  ...rest,
-                  count: Number(count),
-                  image,
-                }
-                acc[topicId].push(res)
-                return acc
-              },
-              {} as GroupedLegislators
-            )
-            return topics.map(
-              ({ speechCount, legislatorCount, id, ...rest }) => ({
-                ...rest,
-                id,
-                speechCount: Number(speechCount),
-                legislatorCount: Number(legislatorCount),
-                legislators: top5legislatorsGrouped[id],
-              })
-            )
-          },
           topNTopicsOfLegislators: async (
             _root,
             { legislatorIds, meetingId, sessionIds, take },
@@ -202,6 +90,7 @@ const extendGraphqlSchema = (baseSchema: GraphQLSchema) => {
         },
       },
       recentSpeechTopicStatsResolvers,
+      topicsOrderBySpeechCountResolver,
     ],
   })
 }
